+++ folio-spring-base/src/main/java/org/folio/spring/config/FolioSpringConfiguration.java
@@ -54,5 +54,4 @@ public class FolioSpringConfiguration {
     @Autowired @Lazy FolioExecutionContext folioExecutionContext) {
     return new DataSourceSchemaAdvisorBeanPostProcessor(folioExecutionContext);
   }
-
 }