+++ folio-spring-base/src/main/java/org/folio/spring/utils/RequestLoggingLevel.java
@@ -1,0 +1,8 @@
+package org.folio.spring.utils;
+
+public enum RequestLoggingLevel {
+  NONE,
+  BASIC,
+  HEADERS,
+  FULL
+}