+++ folio-spring-i18n/src/main/java/org/folio/spring/i18n/model/TranslationFile.java
@@ -1,7 +1,5 @@
 package org.folio.spring.i18n.model;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
@@ -10,6 +8,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.core.io.Resource;
+import tools.jackson.core.type.TypeReference;
+import tools.jackson.databind.ObjectMapper;
 
 /**
  * A class to represent a file that contains translation information.  This file will have a name