+++ folio-spring-system-user/src/main/java/org/folio/spring/client/AuthnClient.java
@@ -2,27 +2,26 @@ package org.folio.spring.client;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.service.annotation.DeleteExchange;
+import org.springframework.web.service.annotation.HttpExchange;
+import org.springframework.web.service.annotation.PostExchange;
 
 @Deprecated(since = "10.0.0", forRemoval = true)
-@FeignClient(name = "folio-spring-base-authn-client", url = "authn")
+@HttpExchange(url = "authn", contentType = APPLICATION_JSON_VALUE)
 public interface AuthnClient {
 
-  @PostMapping(value = "/login-with-expiry", consumes = APPLICATION_JSON_VALUE)
-  ResponseEntity<LoginResponse> loginWithExpiry(@RequestBody UserCredentials credentials);
+  @PostExchange("/login-with-expiry")
+  ResponseEntity<LoginResponse> loginWithExpiry(UserCredentials credentials);
 
-  @PostMapping(value = "/login", consumes = APPLICATION_JSON_VALUE)
-  ResponseEntity<LoginResponse> login(@RequestBody UserCredentials credentials);
+  @PostExchange("/login")
+  ResponseEntity<LoginResponse> login(UserCredentials credentials);
 
-  @PostMapping(value = "/credentials", consumes = APPLICATION_JSON_VALUE)
-  void saveCredentials(@RequestBody UserCredentials credentials);
+  @PostExchange("/credentials")
+  void saveCredentials(UserCredentials credentials);
 
-  @DeleteMapping(value = "/credentials", consumes = APPLICATION_JSON_VALUE)
+  @DeleteExchange("/credentials")
   void deleteCredentials(@RequestParam("userId") String userId);
 
   record UserCredentials(String username, String password) {